Completed saving the results in the csv files as well as displaying results on the GUI and terminal

+++ Projects/P1 Quiz_via_CSV/p1_main.py
@@ -21,9 +21,9 @@ with sqlite3.connect("project1_quiz_cs384.db") as db:
 
 cursor.execute('''
 CREATE TABLE IF NOT EXISTS project1_registration(
-username VARCHAR NOT NULL,
-password VARCHAR NOT NULL,
-name VARCHAR NOT NULL,
+username VARCHAR(8) NOT NULL,
+password VARCHAR(20) NOT NULL,
+name VARCHAR(50) NOT NULL,
 whatsapp number INTEGER(10) NOT NULL
 );
 ''')
@@ -44,10 +44,12 @@ def registration():
     h = hashlib.md5(password.encode())
     db = sqlite3.connect('project1_quiz_cs384.db')
     c = db.cursor()
-    c.execute('DELETE FROM project1_registration WHERE username = (?)',(username))
+    c.execute('''SELECT username AND password FROM project1_registration WHERE username=?;''', (username,))
+    if c.fetchall():
+        c.execute('DELETE FROM project1_registration WHERE username = (?)',(username))
     c.execute("INSERT INTO project1_registration (username,password,name,whatsapp) VALUES (?,?,?,?)",(username,h.hexdigest(),name,whatsapp))
     db.commit()
-    quiz(username,0)
+    quiz(username)
 
 def login():
     user = input('Username: ')
@@ -57,7 +59,7 @@ def login():
     c = db.cursor()
     c.execute('SELECT * FROM project1_registration WHERE username = ? AND password = ?', (user, h.hexdigest()))
     if c.fetchall():
-        quiz(user,1)
+        quiz(user)
     else:
         print('Login failed. Kindly register yourself by filling in the following details')
         registration()
@@ -104,6 +106,95 @@ def make_list_of_q():
             compulsory.append("No")
     os.chdir(cwd)
 
+# saving results 
+def save_results():
+    global tot_marks,correct_q,wrong_q,total,q_no,attempt_q
+    tot_marks=0
+    correct_q=0
+    wrong_q=0
+    i=0
+    skip=0
+    cwd = os.getcwd()
+    path = os.path.join(cwd,f"quiz_wise_questions")
+    os.chdir(path)
+    global text
+    df = pd.read_csv(f"q{n}.csv")
+    total=df['marks_correct_ans'].sum()
+    df=df.drop(df.columns[-1], axis=1)
+    q_no=len(df.index)
+    df['marked_choice']='0'
+    df['Total']='0'
+    df['Legend']='Unattempted'
+    text = df.values.tolist()
+
+    for row in text:
+        if i<len(marked_ans):
+            if marked_ans[i]+1==right_answer[i]:
+                correct_q+=1
+                tot_marks+=positive_marks[i]
+                row[12]='Correct Choices'
+                row[10]=marked_ans[i]+1
+            elif marked_ans[i]==0 or marked_ans[i]==1 or marked_ans[i]==2 or marked_ans[i]==3:
+                wrong_q+=1
+                tot_marks+=neg_marks[i]
+                row[12]='Wrong Choices'
+                row[10]=marked_ans[i]+1
+            elif compulsory[i]=="Yes":
+                tot_marks+=neg_marks[i]
+                skip+=1
+            else:
+                skip+=1
+            row[11]=tot_marks
+            i+=1
+    if i<len(marked_ans)-1:
+        text[i:][11]=tot_marks
+    attempt_q=len(marked_ans)-skip
+    # displaying of results 
+    print("*"*100)
+    print("  ")
+    print(f'Total Quiz Questions: {q_no}')
+    print(f'Total Quiz Questions Attempted: {attempt_q}')
+    print(f'Total Correct Question: {correct_q}')
+    print(f'Total Wrong Questions: {wrong_q}')
+    print(f'Total Marks: {tot_marks}/{total}')
+    os.chdir(cwd)
+    db = sqlite3.connect('project1_quiz_cs384.db')
+    c = db.cursor()
+    c.execute('''SELECT roll AND quiz_num FROM project1_marks WHERE roll=? AND quiz_num = (?)''', (roll,n))
+    if c.fetchall():
+        c.execute('DELETE FROM project1_marks WHERE roll = (?) AND quiz_num = (?)',(roll,n))
+    c.execute("INSERT INTO project1_marks (roll,quiz_num,total_marks) VALUES (?,?,?)",(roll,n,tot_marks))
+    db.commit()
+    path = os.path.join(cwd,"individual_responses")
+    os.chdir(path)
+    if os.path.exists('q'+n+'_'+roll+'.csv'):
+        os.remove('q'+n+'_'+roll+'.csv')
+    with open('q'+n+'_'+roll+'.csv', 'a', newline='') as file:
+        writer = csv.writer(file)
+        writer.writerow(['ques_no','question','option1','option2','option3','option4','correct_option','marks_correct_ans','marks_wrong_ans','compulsory','marked_choice','Total','Legend'])
+        for row in text:
+            writer.writerow(row)
+        writer.writerow(['','','','','','','','','','',tot_marks,'Marks Obtained'])
+        writer.writerow(['','','','','','','','','','',total,'Total Quiz Marks'])
+    os.chdir(cwd)
+
+# displaying results 
+def showresult():
+    lblQuestion.destroy()
+    r1.destroy()
+    r2.destroy()
+    r3.destroy()
+    r4.destroy()
+    r5.destroy()
+    lblRules.destroy()
+    labelresulttext = Label(
+        root,
+        text = "RESULTS\n\nTotal Quiz Questions: "+str(q_no)+"\nTotal Quiz Questions Attempted: "+str(attempt_q)+"\nTotal Correct Question: "+str(correct_q)+"\nTotal Wrong Questions: "+str(wrong_q)+"\nTotal Marks: "+str(tot_marks)+"/"+str(total),
+        font = ("Comic sans MS",20,"bold"),
+        background = "#a3a3c2",
+    )
+    labelresulttext.pack(pady=(150,50))
+
 # taking in user answer 
 def selected():
     global radiovar,user_answer
@@ -114,15 +205,17 @@ def selected():
     radiovar.set(-1)
     if ques < len(questions):
         lblQuestion.config(text= questions[indexes[ques]])
-        lblRules.config(text= "Credits if Correct Option: "+str(positive_marks[ques])+"\nNegative Marking: "+str(neg_marks[ques])+"\nIs compulsory: "+compulsory[ques-1])
+        lblRules.config(text= "Credits if Correct Option: "+str(positive_marks[ques])+"\nNegative Marking: "+str(neg_marks[ques])+"\nIs compulsory: "+compulsory[ques])
         r1['text'] = answers_choice[indexes[ques]][0]
         r2['text'] = answers_choice[indexes[ques]][1]
         r3['text'] = answers_choice[indexes[ques]][2]
         r4['text'] = answers_choice[indexes[ques]][3]
         ques += 1
     else:
+        save_results()
+        showresult()
         # calc()
-        print(marked_ans)
+        # print(marked_ans)
 
 # printing questions and options 
 def startquiz():
@@ -214,8 +307,9 @@ def startIspressed():
     make_list_of_q()
     startquiz()
 
-def quiz(roll,lr):
-    global n,root,labeltext,btnStart
+def quiz(r):
+    global n,root,labeltext,btnStart,roll
+    roll=r
     n=input("Kindly input the quiz set you woulld like to attempt from 1-3:")
 
     root = tkinter.Tk()
